+++ src/Oro/Bundle/OrderBundle/Form/Type/OrderLineItemDraftType.php
@@ -4,15 +4,10 @@ declare(strict_types=1);
 
 namespace Oro\Bundle\OrderBundle\Form\Type;
 
-use Doctrine\Common\Collections\ArrayCollection;
 use Oro\Bundle\FormBundle\Form\Type\OroDateType;
-use Oro\Bundle\FormBundle\Utils\FormUtils;
-use Oro\Bundle\OrderBundle\Entity\Order;
 use Oro\Bundle\OrderBundle\Entity\OrderLineItem;
 use Oro\Bundle\OrderBundle\Provider\OrderLineItemTierPricesProvider;
 use Oro\Bundle\PricingBundle\Entity\PriceTypeAwareInterface;
-use Oro\Bundle\ProductBundle\Entity\Product;
-use Oro\Bundle\ProductBundle\Form\Type\ProductSelectType;
 use Oro\Bundle\ProductBundle\Form\Type\ProductUnitSelectionType;
 use Oro\Bundle\ProductBundle\Form\Type\QuantityType;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
@@ -50,18 +45,7 @@ final class OrderLineItemDraftType extends AbstractType
                     $orderLineItem->setIsFreeForm((bool)(int)$value);
                 },
             ])
-            ->add(
-                'product',
-                ProductSelectType::class,
-                [
-                    'required' => false,
-                    'autocomplete_alias' => 'oro_order_product_visibility_limited',
-                    'grid_name' => 'products-select-grid',
-                    'grid_parameters' => ['types' => [Product::TYPE_SIMPLE, Product::TYPE_KIT]],
-                    'create_enabled' => false,
-                    'data_parameters' => ['scope' => 'order'],
-                ]
-            )
+            ->add('product', OrderLineItemDraftProductType::class)
             ->add('quantity', QuantityType::class, ['required' => true, 'default_data' => 1])
             ->add('productUnit', ProductUnitSelectionType::class, ['required' => true, 'sell' => true])
             ->add(
@@ -83,15 +67,6 @@ final class OrderLineItemDraftType extends AbstractType
             ->addEventListener(FormEvents::POST_SET_DATA, $this->addFreeFormProductOnIsFreeFormPostSetData(...))
             ->addEventListener(FormEvents::POST_SUBMIT, $this->addFreeFormProductOnIsFreeFormPostSubmit(...));
 
-        $builder
-            ->get('product')
-            ->addEventListener(FormEvents::POST_SET_DATA, $this->addKitItemLineItemsOnProductPostSetData(...))
-            ->addEventListener(FormEvents::POST_SET_DATA, $this->replaceProductUnitOnProductPostSetData(...))
-            ->addEventListener(FormEvents::POST_SET_DATA, $this->replacePriceOnProductPostSetData(...))
-            ->addEventListener(FormEvents::POST_SUBMIT, $this->addKitItemLineItemsOnProductPostSubmit(...))
-            ->addEventListener(FormEvents::POST_SUBMIT, $this->replaceProductUnitOnProductPostSubmit(...))
-            ->addEventListener(FormEvents::POST_SUBMIT, $this->replacePriceOnProductPostSubmit(...));
-
         $builder->addEventListener(FormEvents::POST_SUBMIT, $this->fillFreeFormProductOnPostSubmit(...));
 
         $builder->addEventSubscriber($this->orderLineItemDraftDrySubmitListener);
@@ -111,96 +86,6 @@ final class OrderLineItemDraftType extends AbstractType
         }
     }
 
-    private function addKitItemLineItemsOnProductPostSetData(FormEvent $event): void
-    {
-        $form = $event->getForm();
-        /** @var OrderLineItem $orderLineItem */
-        $orderLineItem = $form->getParent()->getData();
-        /** @var Product|null $product */
-        $product = $event->getData();
-        if ($product !== null) {
-            // Checking if the line item is a kit based on the product type.
-            $isKit = $product->isKit();
-        } elseif ($orderLineItem->isFreeForm()) {
-            // Free-form line item cannot be a kit.
-            $isKit = false;
-        } else {
-            // Determines whether the line item is a kit based on the presence of kit item line items.
-            $isKit = $orderLineItem->getKitItemLineItems()->count() > 0;
-        }
-
-        if ($isKit) {
-            /** @var Order $order */
-            $order = $orderLineItem->getOrder();
-            $currency = $order->getCurrency();
-
-            $form->getParent()->add(
-                'kitItemLineItems',
-                OrderProductKitItemLineItemCollectionType::class,
-                [
-                    'required' => true,
-                    'product' => $product,
-                    'currency' => $currency,
-                ]
-            );
-        }
-    }
-
-    private function replaceProductUnitOnProductPostSetData(FormEvent $event): void
-    {
-        /** @var Product|null $product */
-        $product = $event->getData();
-        if (!$product) {
-            return;
-        }
-
-        $form = $event->getForm();
-
-        // FormUtils::replaceField is not used on purpose as it prevents the initialization of new choices.
-        $form->getParent()->add(
-            'productUnit',
-            ProductUnitSelectionType::class,
-            [
-                'required' => true,
-                'product' => $product,
-                'init_choices' => true,
-                'auto_initialize' => false,
-                'empty_data' => $product->getPrimaryUnitPrecision()?->getProductUnitCode(),
-                'sell' => true,
-            ]
-        );
-    }
-
-    private function replacePriceOnProductPostSetData(FormEvent $event): void
-    {
-        $form = $event->getForm();
-        /** @var Product|null $product */
-        $product = $event->getData();
-        $isKit = (bool)$product?->isKit();
-        /** @var OrderLineItem $orderLineItem */
-        $orderLineItem = $form->getParent()->getData();
-        /** @var Order $order */
-        $order = $orderLineItem->getOrder();
-        $currency = $order->getCurrency();
-
-        FormUtils::replaceField(
-            $form->getParent(),
-            'price',
-            [
-                'default_currency' => $currency,
-                'readonly' => $isKit,
-            ]
-        );
-
-        FormUtils::replaceField(
-            $form->getParent()->get('price'),
-            'is_price_changed',
-            [
-                'data' => $orderLineItem->getId() ? '1' : '0'
-            ]
-        );
-    }
-
     private function addFreeFormProductOnIsFreeFormPostSubmit(FormEvent $event): void
     {
         $form = $event->getForm();
@@ -217,84 +102,6 @@ final class OrderLineItemDraftType extends AbstractType
                 ->remove('productSku')
                 ->remove('freeFormProduct');
         }
-    }
-
-    private function addKitItemLineItemsOnProductPostSubmit(FormEvent $event): void
-    {
-        $form = $event->getForm();
-        /** @var Product|null $product */
-        $product = $form->getData();
-        $isKit = (bool)$product?->isKit();
-        /** @var OrderLineItem $orderLineItem */
-        $orderLineItem = $form->getParent()->getData();
-
-        if ($isKit) {
-            /** @var Order $order */
-            $order = $orderLineItem->getOrder();
-            $currency = $order->getCurrency();
-
-            if ($orderLineItem->getProduct() !== $product) {
-                // Kit items collection should be cleared and totally replaced if the product is changed.
-                $orderLineItem->getKitItemLineItems()->clear();
-                $orderLineItem->setKitItemLineItems(new ArrayCollection());
-            }
-
-            $form->getParent()->add(
-                'kitItemLineItems',
-                OrderProductKitItemLineItemCollectionType::class,
-                [
-                    'required' => true,
-                    'product' => $product,
-                    'currency' => $currency,
-                ]
-            );
-        } else {
-            $form->getParent()->remove('kitItemLineItems');
-            $orderLineItem->getKitItemLineItems()->clear();
-        }
-    }
-
-    private function replaceProductUnitOnProductPostSubmit(FormEvent $event): void
-    {
-        $form = $event->getForm();
-        /** @var Product|null $product */
-        $product = $form->getData();
-
-        // FormUtils::replaceField is not used on purpose as it prevents the initialization of new choices.
-        $form->getParent()->add(
-            'productUnit',
-            ProductUnitSelectionType::class,
-            [
-                'required' => true,
-                'product' => $product,
-                'init_choices' => true,
-                'auto_initialize' => false,
-                'empty_data' => $product?->getPrimaryUnitPrecision()?->getProductUnitCode(),
-                'sell' => true,
-            ]
-        );
-    }
-
-    private function replacePriceOnProductPostSubmit(FormEvent $event): void
-    {
-        $form = $event->getForm();
-        /** @var Product|null $product */
-        $product = $form->getData();
-        $isKit = (bool)$product?->isKit();
-        /** @var OrderLineItem $orderLineItem */
-        $orderLineItem = $form->getParent()->getData();
-        /** @var Order $order */
-        $order = $orderLineItem->getOrder();
-        $currency = $order->getCurrency();
-
-        FormUtils::replaceField(
-            $form->getParent(),
-            'price',
-            [
-                'default_currency' => $currency,
-                'readonly' => $isKit,
-            ]
-        );
     }
 
     private function fillFreeFormProductOnPostSubmit(FormEvent $event): void